2022_08_17_Update
Cleaning of the Vignette Markdown with systematization of a readable and understandable Content Table.
Need to be checked for error and Knit into html -> we must add in the introduction of the markdown smth to have the content table in the knitted html.
It should not pop uo errors though because they have been checked in one of the 2022_08_17_Update and the change were only esthetic.